bug: 开源版质量红线接口调整 #5886
